test: WorkflowEngine 統合テストを heavy IT へ分類移動する
